+++ NEWS.md
@@ -4,6 +4,7 @@
 * Added the possibility of multiple group comparisons to `flex_table1()` (recommended by Rémi and Taylor)
 * Added the function `pretty_cm()` for plotting of caret::confusionMatrix()
 * Added the function `get_ICD_10_cats()` that returns ICD-10 F Diagnoses categories (thanks to Dirk Edelbuettel)
+* Added the convenience function `independent_sample_means()` to compare means of 2 or more groups (including check of assumptions).
 
 # datscience 0.2.2